ud-useradd: support usergroups
Patch from HE* that implements per-user groups for ud-useradd.  If run
without -n the behaviour is as before.  With -n (for no-automatic-IDs)
the user gets the chance to change the assigned numeric UID.  The user
is also prompted for a GID.  By default the default group for users is
still used but ud-useradd suggests a free group ID for the user group.
If the user group ID is chosen a group is automatically created.
 (* based on work by aba and joerg)