Two problems with my first commit:
We need to specify the RFC 2047 character set as utf-8, not us-ascii,
for the encoding type field in the header

We need to only encode the FullName, and then tack on the emailaddr.
The examples in the RFC show this behavior, and at least mutt won't
decode the header if the email address is also RFC 2047 encoded (which
makes some sense - currently email addresses can only really be ascii).