Add sshdistAuthKeysHost
We autogenerate the authorized_keys files for sshdist on db-master.  It
limits the hosts' ssh key to coming from their respective addresses.
Now we can add additional source addresses to accept for this since not
all hosts appear to come from their published address (or have a
published address for that matter).