ud-mailgate: fix gpg result usage
We use the result of the pgp check for quite a long time in the main
program.  Give it its own variable instead of using Res which was
overwritten a bit later.  Also make a new gpgcheck2 class that allows us
to access the values of the gpg signature check in a saner way.