ud-generate speed, I
cut down on calls to IsInGroup by doing it once in generate_host() and
not having the individual generators run it.

side effect: Up until now we exported empty groups to a host, if that group had
a user with that group as their primary group - even if that particular user
was not exported to this this.  No we no longer export empty groups.